Move a lot of stuff to the <DesktopGrid>
