+++ src/components/Layout/Layout.module.css
@@ -1,38 +1,3 @@
-@keyframes background {
-  0% {
-    background-color: #ffffff;
-  }
-  20% {
-    background-color: #e3ffb5;
-  }
-  40% {
-    background-color: #e57200;
-  }
-  60% {
-    background-color: #5dade2;
-  }
-  80% {
-    background-color: #cc8899;
-  }
-  100% {
-    background-color: #ffffff;
-  }
-}
-
-.wrapper {
-  font-family: "General Sans", sans-serif;
-  min-height: 100vh;
-  position: relative;
-  background-color: var(--layout-bg-color);
-  animation-name: none;
-  animation-duration: 10s;
-  animation-timing-function: linear;
-  animation-iteration-count: infinite;
-  animation-play-state: running;
-
-  transition: all 0.2s;
-}
-
 .logo {
   font-size: 3.75rem;
   margin: 2rem auto 0;
@@ -62,45 +27,4 @@
   line-height: 2.25rem;
   padding-top: 6.25rem;
   text-align: center;
-}
-
-.bottomWrapper {
-  bottom: 3.75rem;
-  color: #000000;
-  font-size: 2.2rem;
-  left: 0;
-  position: fixed;
-  text-align: center;
-  width: 100%;
-}
-
-.bigLink {
-  color: #000000;
-  text-decoration: none;
-  font-size: 2.2rem;
-
-  &:hover {
-    text-decoration: line-through;
-  }
-}
-
-.fixedLink {
-  position: fixed;
-  z-index: 100;
-}
-
-.top {
-  top: 3.75rem;
-}
-
-.bottom {
-  bottom: 3.75rem;
-}
-
-.left {
-  left: 5rem;
-}
-
-.right {
-  right: 5rem;
 }
